unit-requirements.txt
databricks-feature-store==0.12.1
pyspark==3.3.1
pytest==7.2.1
python-dateutil==2.8.2
pylint==2.16.2
pyyaml==6.0
mlflow
pandas
numpy
scipy
lifetimes==0.10.1
scikit-learn
typing_extensions==4.5.0